refactor: Consolidate version-derived service creation in Bootstrap
